# Conflicts:
#	GenerateJSCode.php
#	INTER-Mediator-Calc.js
#	INTER-Mediator-Context.js
#	INTER-Mediator-Element.js
#	INTER-Mediator-Lib.js
#	INTER-Mediator-UI.js
#	INTER-Mediator-UnitTest/INTER-Mediator-Lib-test.js
#	INTER-Mediator.js
#	Samples/Sample_forDebugging/formattest.html
#	Samples/Sample_forDebugging/formattest.php
#	dist-docs/TestDB.fmp12
#	dist-docs/TestDB.fp7
#	dist-docs/TestDBSchemeForFMS.xlsx
#	dist-docs/TestDB_clone.fmp12
#	params.php

+++ dist-docs/buildup.sh
@@ -1,12 +1,14 @@
 #!/bin/sh
 
 # INTER-Mediator Deployment File Set Builder
-# by Masayuki Nii  msyk@msyk.net Copyright (c) 2010-2015 Masayuki Nii, All rights reserved.
-
-# This project started at the end of 2009.
+# Copyright (c) INTER-Mediator Directive Committee (http://inter-mediator.org)
+# This project started at the end of 2009 by Masayuki Nii msyk@msyk.net.
+#
 # INTER-Mediator is supplied under MIT License.
+# Please see the full license for details:
+# https://github.com/INTER-Mediator/INTER-Mediator/blob/master/dist-docs/License.txt
 
-version="5.1-dev"
+version="5.6-RC2"
 
 # The jar file of YUI can be donwloaded from below.
 # http://grepcode.com/snapshot/repo1.maven.org/maven2/com.yahoo.platform.yui/yuicompressor/2.4.7
@@ -22,15 +24,10 @@ echo " Start to build the INTER-Mediator Ver.${version}"
 echo "-------------------------------------------------"
 
 dt=$(date "+%Y-%m-%d")
-
-sedrule="/tmp/sedrule"
-cat << EOF > "${sedrule}"
-s/@@@@1@@@@/${dt}/
-s/@@@@2@@@@/${version}/
-EOF
-
 distDocDir=$(cd $(dirname "$0"); pwd)
 originalPath=$(dirname "${distDocDir}")
+
+printf '{"version":"%s","releasedate":"%s"}' "${version}" "${dt}" > "${originalPath}/metadata.json"
 
 topOfDir=$(dirname "${originalPath}")
 buildDir="${topOfDir}/${buildRootName}"
@@ -42,11 +39,16 @@ echo " Build to: ${buildPath}"
 echo "-------------------------------------------------"
 echo "Choose the build result from these:"
 echo ' (1) Complete (everything contains)'
-echo ' (2) Core only (the least set to work wep applications)'
+echo ' (2) Core only (the least set to work web applications)'
 echo ' (3) Core + Support (add Auth_Support and INTER-Mediator-Support)'
-/bin/echo -n "Type 1, 2 or 3, and then type return----> "
+echo ' (4) Write just version and release date to metadata.json'
+/bin/echo -n "Type 1, 2, 3 or 4, and then type return----> "
 read choice
 echo ""
+
+if [ $choice = 4 ]; then
+    exit 0;
+fi
 
 if [ -d "${buildDir}" ]; then
     rm -r "${buildDir}"
@@ -58,13 +60,14 @@ cd "${originalPath}"
 for aFile in $(ls *.php)
 do
     filename=$(basename "${aFile}")
-#    echo "SED: ${aFile}"
-    sed -f "${sedrule}" "${aFile}" > "${buildPath}/${filename}"
+    cp "${aFile}" "${buildPath}/${filename}"
 done
+
+cp  "${originalPath}/metadata.json" "${buildPath}/metadata.json"
 
 #### Merge js files
 echo "PROCESSING: Merging JS files"
-sed -f "${sedrule}" "${originalPath}/INTER-Mediator.js"        > "${buildPath}/temp.js"
+cp  "${originalPath}/INTER-Mediator.js"                          "${buildPath}/temp.js"
 cat "${originalPath}/INTER-Mediator-Element.js"               >> "${buildPath}/temp.js"
 cat "${originalPath}/INTER-Mediator-Events.js"                >> "${buildPath}/temp.js"
 cat "${originalPath}/INTER-Mediator-Context.js"               >> "${buildPath}/temp.js"
@@ -75,46 +78,62 @@ cat "${originalPath}/INTER-Mediator-Page.js"                  >> "${buildPath}/t
 cat "${originalPath}/INTER-Mediator-Parts.js"                 >> "${buildPath}/temp.js"
 cat "${originalPath}/INTER-Mediator-Navi.js"                  >> "${buildPath}/temp.js"
 cat "${originalPath}/INTER-Mediator-UI.js"                    >> "${buildPath}/temp.js"
-cat "${originalPath}/lib/js_lib/tinySHA1.js"                  >> "${buildPath}/temp.js"
-cat "${originalPath}/lib/js_lib/sha256.js"                    >> "${buildPath}/temp.js"
+if [ ! -f "${topOfDir}/${YUICOMP}" ]; then
+    cat "${originalPath}/lib/js_lib/tinySHA1.js"              >> "${buildPath}/temp.js"
+    echo ';'                                                  >> "${buildPath}/temp.js"
+    cat "${originalPath}/lib/js_lib/sha256.js"                >> "${buildPath}/temp.js"
+fi
 cat "${originalPath}/lib/bi2php/biBigInt.js"                  >> "${buildPath}/temp.js"
 cat "${originalPath}/lib/bi2php/biMontgomery.js"              >> "${buildPath}/temp.js"
 cat "${originalPath}/lib/bi2php/biRSA.js"                     >> "${buildPath}/temp.js"
 cat "${originalPath}/Adapter_DBServer.js"                     >> "${buildPath}/temp.js"
 cat "${originalPath}/INTER-Mediator-DoOnStart.js"             >> "${buildPath}/temp.js"
 
-sed -f "${sedrule}" "${buildPath}/temp.js" > "${buildPath}/INTER-Mediator.js"
+cp "${buildPath}/temp.js" "${buildPath}/INTER-Mediator.js"
 
 #### Compress INTER-Mediator.js
 if [ -f "${topOfDir}/${YUICOMP}" ]; then
     sed '1s/*/*!/' "${buildPath}/temp.js" > "${buildPath}/temp2.js"
-    
+
     osName=$(uname -s)
     echo "Detected OS: ${osName}"
     if [[ "${osName}" == CYGWIN* ]];  then
-    	jarPath=$(cygpath -w "${topOfDir}/${YUICOMP}")
-    	temp2Path=$(cygpath -w "${buildPath}/temp2.js")
-    	temp3Path=$(cygpath -w "${buildPath}/temp3.js")
-    	yuiLogPath=$(cygpath -w "${buildDir}/${YUICOMPLOG}")
-    else 
-    	jarPath="${topOfDir}/${YUICOMP}"
-    	temp2Path="${buildPath}/temp2.js"
-    	temp3Path="${buildPath}/temp3.js"
-    	yuiLogPath="${buildDir}/${YUICOMPLOG}"
+        jarPath=$(cygpath -w "${topOfDir}/${YUICOMP}")
+        temp2Path=$(cygpath -w "${buildPath}/temp2.js")
+        temp3Path=$(cygpath -w "${buildPath}/temp3.js")
+        yuiLogPath=$(cygpath -w "${buildDir}/${YUICOMPLOG}")
+    else
+        jarPath="${topOfDir}/${YUICOMP}"
+        temp2Path="${buildPath}/temp2.js"
+        temp3Path="${buildPath}/temp3.js"
+        yuiLogPath="${buildDir}/${YUICOMPLOG}"
     fi
     java -jar "${jarPath}"  "${temp2Path}" -v --charset UTF-8 -o "${temp3Path}" 2> "${yuiLogPath}"
     sed '1s/*!/*/' "${temp3Path}" > "${buildPath}/INTER-Mediator.js"
+    head -n 9 "${buildPath}/INTER-Mediator.js"           > "${buildPath}/temp.js"
+    tail -n 1 "${originalPath}/lib/js_lib/tinySHA1.js"  >> "${buildPath}/temp.js"
+    echo ';'                                            >> "${buildPath}/temp.js"
+    tail -n 1 "${originalPath}/lib/js_lib/sha256.js"    >> "${buildPath}/temp.js"
+    tail -n 1 "${buildPath}/INTER-Mediator.js"          >> "${buildPath}/temp.js"
+    mv "${buildPath}/temp.js" "${buildPath}/INTER-Mediator.js"
     rm  "${buildPath}/temp.js" "${temp2Path}" "${temp3Path}"
 else
     rm  "${buildPath}/temp.js"
 fi
 
+# Copy "DB_Support" directory.
+echo "PROCESSING: ${originalPath}/DB_Support"
+cp -prf "${originalPath}/DB_Support" "${buildPath}"
+
 # Copy "lib" path php contents.
 echo "PROCESSING: ${originalPath}/lib"
 mkdir -p "${buildPath}/lib/bi2php"
 cp -p "${originalPath}/lib/bi2php/biRSA.php" "${buildPath}/lib/bi2php"
+cp -prf "${originalPath}/lib/CWPKit" "${buildPath}/lib"
 cp -prf "${originalPath}/lib/FX" "${buildPath}/lib"
-cp -prf "${originalPath}/lib/phpseclib" "${buildPath}/lib"
+cp -prf "${originalPath}/lib/ParagonIE" "${buildPath}/lib"
+cp -prf "${originalPath}/lib/phpseclib_v1" "${buildPath}/lib"
+cp -prf "${originalPath}/lib/phpseclib_v2" "${buildPath}/lib"
 cp -prf "${originalPath}/lib/mailsend" "${buildPath}/lib"
 
 if [ $choice = 3 ]; then
@@ -133,8 +152,7 @@ do
     do
 #        echo "Processing: ${originalPath}/${TARGET}/${DIR}"
         if [ -f "${originalPath}/${TARGET}/${DIR}" ]; then
-#            echo "SED: ${originalPath}/${TARGET}/${DIR}"
-            sed -f "${sedrule}" "${originalPath}/${TARGET}/${DIR}" > "${buildPath}/${TARGET}/${DIR}"
+            cp "${originalPath}/${TARGET}/${DIR}" "${buildPath}/${TARGET}/${DIR}"
         else
             mkdir -p "${buildPath}/${TARGET}/${DIR}"
             for FILE in $(ls "${originalPath}/${TARGET}/${DIR}")
@@ -143,7 +161,7 @@ do
                     case "${FILE}" in
                         *\.html | *\.php | *\.js | *\.css | *\.txt)
 #                          echo "SED: ${originalPath}/${TARGET}/${DIR}/${FILE}"
-                            sed -f "${sedrule}" "${originalPath}/${TARGET}/${DIR}/${FILE}" > "${buildPath}/${TARGET}/${DIR}/${FILE}"
+                            cp "${originalPath}/${TARGET}/${DIR}/${FILE}" "${buildPath}/${TARGET}/${DIR}/${FILE}"
                             ;;
                         *)
 #                            echo "CP: ${originalPath}/${TARGET}/${DIR}/${FILE}"
@@ -165,25 +183,25 @@ if [ $choice = 1 ]; then
 
     echo "PROCESSING: Rest of ${originalPath}/Samples"
     cp -pr  "${originalPath}/Samples/Sample_products/images" "${buildPath}/Samples/Sample_products/"
-    cp -pr  "${originalPath}/Samples/WebSite/previous_rsrcs" "${buildPath}/Samples/WebSite/"
 
 # Invalidate the definition file of the DefEditor.
-    echo "PROCESSING: Invalidate the Definition File Editor for security reason."
-    defeditdeffile="${buildPath}/INTER-Mediator-Support/defedit.php"
-    sed 's|IM_Entry|/* IM_Entry|' "${defeditdeffile}" > /tmp/defedit.php
-    cp -p /tmp/defedit.php "${defeditdeffile}"
+#    echo "PROCESSING: Invalidate the Definition File Editor for security reason."
+#    defeditdeffile="${buildPath}/INTER-Mediator-Support/defedit.php"
+#    sed 's|IM_Entry|/* IM_Entry|' "${defeditdeffile}" > /tmp/defedit.php
+#    cp -p /tmp/defedit.php "${defeditdeffile}"
 else
     echo "PROCESSING: ${originalPath}/dist-docs/License.txt"
     cp -p   "${originalPath}/dist-docs/License.txt" "${buildPath}"
-	readmeLines=`wc -l "${originalPath}/dist-docs/readme.txt" | awk '{print $1}'`
-	lines=`expr $readmeLines - 8`
+    readmeLines=`wc -l "${originalPath}/dist-docs/readme.txt" | awk '{print $1}'`
+    lines=`expr $readmeLines - 8`
     echo "PROCESSING: ${originalPath}/dist-docs/readme.txt"
     head -n `echo $lines` "${originalPath}/dist-docs/readme.txt" > "${buildPath}/readme.txt"
 fi
 
-find "${buildPath}" -name "\.*" -exec rm -rf {} \;
+echo "PROCESSING: ${originalPath}/themes"
+cp -prf "${originalPath}/themes" "${buildPath}"
 
-rm "${sedrule}"
+find "${buildPath}" -name "\.*" -exec rm -rf {} \;
 
 #
 echo ""
@@ -198,7 +216,7 @@ echo "Build to: ${buildPath}" >> "${buildDir}/${receipt}"
 if [ $choice = 1 ]; then
     echo 'Your Choice: (1) Complete (everything contains)' >> "${buildDir}/${receipt}"
 elif [ $choice = 2 ]; then
-    echo 'Your Choice: (2) Core only (the least set to work wep applications)' >> "${buildDir}/${receipt}"
+    echo 'Your Choice: (2) Core only (the least set to work web applications)' >> "${buildDir}/${receipt}"
 elif [ $choice = 3 ]; then
     echo 'Your Choice: (3) Core + Support (add Auth_Support and INTER-Mediator-Support)' >> "${buildDir}/${receipt}"
 else
